Split off param creation
